Result from evening hack
The language consists of AST data structures and an interpreter. The
interpreter has a few built in functions. The main program is hard-coded
to run an "echo" program.

Maskinspraak8000.Interpreter.insertVal is where the magic happens.
Environments and functions refer to each other in a cyclic way. No
mutation is needed!